+++ SAGESharp/SAGESharp.csproj
@@ -57,13 +57,13 @@
     <Compile Include="SLB\Identifier.cs" />
     <Compile Include="SLB\IdentifierBinaryReader.cs" />
     <Compile Include="SLB\Level\Conversation\Character.cs" />
-    <Compile Include="SLB\Level\Conversation\LineSide.cs" />
+    <Compile Include="SLB\Level\Conversation\CharacterBinaryReader.cs" />
+    <Compile Include="SLB\Level\Conversation\ConversationBinaryReader.cs" />
     <Compile Include="SLB\Level\Conversation\Frame.cs" />
+    <Compile Include="SLB\Level\Conversation\FrameBinaryReader.cs" />
     <Compile Include="SLB\Level\Conversation\Info.cs" />
-    <Compile Include="SLB\Level\Conversation\Internal\CharacterBinaryReader.cs" />
-    <Compile Include="SLB\Level\Conversation\Internal\ConversationBinaryReader.cs" />
-    <Compile Include="SLB\Level\Conversation\Internal\FrameBinaryReader.cs" />
-    <Compile Include="SLB\Level\Conversation\Internal\InfoBinaryReader.cs" />
+    <Compile Include="SLB\Level\Conversation\InfoBinaryReader.cs" />
+    <Compile Include="SLB\Level\Conversation\LineSide.cs" />
     <Compile Include="SLB\SLBBinaryReaderFactory.cs" />
     <Compile Include="SLB\StreamExtensions.cs" />
     <Compile Include="SLB\UIntExtensions.cs" />
