+++ SAGESharp/SLB/Level/Conversation/FrameBinaryReader.cs
@@ -2,7 +2,7 @@
 using System.IO;
 using System.Text;
 
-namespace SAGESharp.SLB.Level.Conversation.Internal
+namespace SAGESharp.SLB.Level.Conversation
 {
     /// <summary>
     /// Class to read Frame objects from binary SLB files.
