+++ SAGESharpTests/SAGESharpTests.csproj
@@ -66,14 +66,14 @@
     <Compile Include="SLB\IdentifierEqualityTests.cs" />
     <Compile Include="SLB\IdentifierTests.cs" />
     <Compile Include="SLB\IdentifierBinaryReaderTests.cs" />
+    <Compile Include="SLB\Level\Conversation\CharacterBinaryReaderTests.cs" />
     <Compile Include="SLB\Level\Conversation\CharacterEqualityTests.cs" />
+    <Compile Include="SLB\Level\Conversation\ConversationBinaryReaderTests.cs" />
     <Compile Include="SLB\Level\Conversation\Defaults.cs" />
+    <Compile Include="SLB\Level\Conversation\FrameBinaryReaderTests.cs" />
     <Compile Include="SLB\Level\Conversation\FrameEqualityTests.cs" />
+    <Compile Include="SLB\Level\Conversation\InfoBinaryReaderTests.cs" />
     <Compile Include="SLB\Level\Conversation\InfoEqualityTests.cs" />
-    <Compile Include="SLB\Level\Conversation\Internal\CharacterBinaryReaderTests.cs" />
-    <Compile Include="SLB\Level\Conversation\Internal\ConversationBinaryReaderTests.cs" />
-    <Compile Include="SLB\Level\Conversation\Internal\FrameBinaryReaderTests.cs" />
-    <Compile Include="SLB\Level\Conversation\Internal\InfoBinaryReaderTests.cs" />
     <Compile Include="SLB\StreamExtensionsTests.cs" />
     <Compile Include="SLB\UIntExtensionsTests.cs" />
     <Compile Include="Util\AbstractEqualityByRefTests.cs" />
