+++ SAGESharpTests/SLB/Level/Conversation/CharacterBinaryReaderTests.cs
@@ -2,12 +2,11 @@
 using NUnit.Framework;
 using SAGESharp.SLB;
 using SAGESharp.SLB.Level.Conversation;
-using SAGESharp.SLB.Level.Conversation.Internal;
 using SAGESharpTests.Util;
 using System;
 using System.IO;
 
-namespace SAGESharpTests.SLB.Level.Conversation.Internal
+namespace SAGESharpTests.SLB.Level.Conversation
 {
     [TestFixture]
     public static class CharacterBinaryReaderTests
